pyproject.toml
[project]
name = "drone-agent"
version = "0.1.0"
description = "Drone Agent - Drone autonomy using Generative AI"
requires-python = ">=3.10"
dependencies = [
    "djitellopy2~=2.3",
    "google-genai>=1.17.0",
    "langfuse>=2.60.7",
    "mss>=10.0.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "smolagents[all]>=1.17.0",
    "tenacity>=9.1.2"
]
